chore: move source files to src, split code
