+++ src/utils/crypto.ts
@@ -1,15 +1,12 @@
 import { KeyPair } from "libsodium-wrappers";
 const _sodium = require("libsodium-wrappers");
 
-const PH_VERSION = "v1";
-type PhaseCiphertext = `ph:${string}:${string}:${string}:${string}`;
-
 /**
  * Returns an ephemeral key exchange keypair
  *
  * @returns {KeyPair}
  */
-const randomKeyPair = async () => {
+export const randomKeyPair = async () => {
   await _sodium.ready;
   const sodium = _sodium;
   const keypair = await sodium.crypto_kx_keypair();
@@ -24,7 +21,7 @@ const randomKeyPair = async () => {
  * @param {Uint8Array} recipientPubKey
  * @returns
  */
-const sessionKeys = async (
+export const sessionKeys = async (
   ephemeralKeyPair: KeyPair,
   recipientPubKey: Uint8Array
 ) => {
@@ -68,56 +65,9 @@ const encryptRaw = async (plaintext: String, key: Uint8Array) => {
  * @param {Uint8Array} key - Symmetric encryption key
  * @returns {string}
  */
-const encryptString = async (plaintext: string, key: Uint8Array) => {
+export const encryptString = async (plaintext: string, key: Uint8Array) => {
   await _sodium.ready;
   const sodium = _sodium;
 
   return sodium.to_hex(await encryptRaw(sodium.from_string(plaintext), key));
 };
-
-export default class Phase {
-  appPubKey: string;
-
-  constructor(appId: string) {
-    const appIdRegex = /^phApp:v(\d+):([a-fA-F0-9]{64})$/;
-    if (!appIdRegex.test(appId)) {
-      throw new Error("Invalid Phase appID");
-    }
-    this.appPubKey = appId.split(":")[2];
-  }
-
-  encrypt = async (
-    plaintext: string,
-    tag: string = ""
-  ): Promise<PhaseCiphertext> => {
-    await _sodium.ready;
-    const sodium = _sodium;
-
-    return new Promise<PhaseCiphertext>(async (resolve, reject) => {
-      try {
-        const oneTimeKeyPair = await randomKeyPair();
-        const symmetricKeys = await sessionKeys(
-          oneTimeKeyPair,
-          sodium.from_hex(this.appPubKey)
-        );
-        const ciphertext = await encryptString(
-          plaintext,
-          symmetricKeys.sharedTx
-        );
-
-        // Use sodium.memzero to wipe the keys from memory
-        sodium.memzero(oneTimeKeyPair.privateKey);
-        sodium.memzero(symmetricKeys.sharedTx);
-        sodium.memzero(symmetricKeys.sharedRx);
-
-        resolve(
-          `ph:${PH_VERSION}:${sodium.to_hex(
-            oneTimeKeyPair.publicKey
-          )}:${ciphertext}:${tag}`
-        );
-      } catch (error) {
-        reject(`Something went wrong: ${error}`);
-      }
-    });
-  };
-}
